population api

+++ app/models.py
@@ -2,15 +2,23 @@
 
 import pandas as pd
 import json
+import os
 
 class Population():
     def initialize_data(csv_file):
-        df = pd.read_csv(csv_file, usecols=['Country name', 'Year', 'Population'])
-        db = {}
-        for index, row in df.iterrows():
-            country_name = row['Country name']
-            year = row['Year']
-            population = row['Population']
-            db.setdefault(year, []).append((country_name, population))
-        fh = open("db.json", 'w')
-        json.dump(db, fh)
+        if os.path.exists("db.json"):
+            print("Database file already exists. Skipping initialization.")
+            return
+        
+        try:
+            df = pd.read_csv(csv_file, usecols=['Country name', 'Year', 'Population'])
+            db = {}
+            for index, row in df.iterrows():
+                country_name = row['Country name']
+                year = row['Year']
+                population = row['Population']
+                db.setdefault(year, []).append((country_name, population))
+            with open("db.json", 'w') as fh:
+                json.dump(db, fh)
+        except Exception as e:
+            print(f"Error: {e}")
